refactor(design): colocate prototype state catalogs with components
Move each component's variant catalog out of the prototype's monolithic
states.tsx into a sibling *Configurations.tsx next to the component.
Prototype state.tsx files shrink to thin assembly scripts that import
each component's configurations array.

For shared forms used by multiple prototypes (ProfileDetailsForm,
AddressForm, PaymentMethodForm), the Configurations file exposes
multiple named arrays (e.g. addressFormManagementConfigurations,
addressFormSelfOnboardingConfigurations) so each prototype picks the
variant set it needs.

Each component now has a *States.tsx (demo adapter) + *Configurations.tsx
(variant catalog) pair. The split keeps Vite's react-refresh HMR working
cleanly — files exporting both a component and non-component data break
fast-refresh.
