+++ sdk-app/src/design/components/contractor/self-onboarding/ContractorOnboardingComplete/ContractorOnboardingCompleteStates.tsx
@@ -1,0 +1,9 @@
+import { ContractorOnboardingComplete } from './ContractorOnboardingComplete'
+
+/**
+ * Renders ContractorOnboardingComplete for state demos with a no-op
+ * Done handler.
+ */
+export function ContractorOnboardingCompleteDemo() {
+  return <ContractorOnboardingComplete onDone={() => {}} />
+}
