+++ sdk-app/src/design/components/contractor/shared/PaymentMethodForm/PaymentMethodFormStates.tsx
@@ -1,0 +1,14 @@
+import { PaymentMethodForm, type PaymentMethodFormDefaults } from './PaymentMethodForm'
+
+export interface PaymentMethodFormDemoProps {
+  heading: string
+  description: string
+  defaultValues: PaymentMethodFormDefaults
+}
+
+/**
+ * Renders PaymentMethodForm for state demos with a no-op submit handler.
+ */
+export function PaymentMethodFormDemo(props: PaymentMethodFormDemoProps) {
+  return <PaymentMethodForm {...props} onSubmit={async () => Promise.resolve()} />
+}
